pyproject.toml
[project]
name = "microgpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-neo4j>=0.9.0",
    "neo4j>=6.1.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "sentence-transformers>=5.4.1",
]
